Move error type-to-message conversion to parent activity
